feat: clean up lace-imports re-exports - move types to @/types/core
MAIN CHANGES:
- Remove `export * from '@/types/core'` re-export from lace-imports.ts (eliminates maze)
- Move EVENT_TYPES and ThreadEventType to packages/web/types/core.ts
- Remove EVENT_TYPES/ThreadEventType exports from lace-imports.ts
- Update 8 key files to import core types from @/types/core instead of lace-imports

lace-imports.ts now only exports business logic classes, not types.
All type imports should go directly to @/types/core.

Remaining files will be updated in follow-up commits.
