+++ packages/web/components/timeline/tool/types.ts
@@ -1,7 +1,7 @@
 // ABOUTME: Type definitions for tool renderer system
 // ABOUTME: Defines interfaces and helper types for customizable tool display logic
 
-import type { ToolResult } from '@/lib/server/lace-imports';
+import type { ToolResult } from '@/types/core';
 import type { ToolAggregatedEventData } from '@/types/web-events';
 
 export interface ToolRenderer {
