+++ packages/web/lib/server/lace-imports.ts
@@ -12,7 +12,3 @@ export { EventApprovalCallback } from '~/tools/event-approval-callback';
 export { Session } from '~/sessions/session';
 export { Project } from '~/projects/project';
 export { ThreadManager } from '~/threads/thread-manager';
-export { EVENT_TYPES, type ThreadEventType } from '~/threads/types';
-
-// Re-export types and utilities from unified core types
-export * from '@/types/core';
